Moved TaskSpecResponseDTO to arrebol package
