+++ src/main/java/org/fogbowcloud/saps/engine/core/scheduler/executor/arrebol/dtos/TaskResponseDTO.java
@@ -3,7 +3,6 @@ package org.fogbowcloud.saps.engine.core.scheduler.executor.arrebol.dtos;
 import java.io.Serializable;
 
 import com.google.gson.annotations.SerializedName;
-import org.fogbowcloud.saps.engine.core.dto.TaskSpecResponseDTO;
 
 public class TaskResponseDTO implements Serializable {
     /**
