+++ src/main/java/org/fogbowcloud/saps/engine/core/scheduler/executor/arrebol/dtos/TaskSpecResponseDTO.java
@@ -1,8 +1,9 @@
-package org.fogbowcloud.saps.engine.core.dto;
+package org.fogbowcloud.saps.engine.core.scheduler.executor.arrebol.dtos;
 
 import java.io.Serializable;
 import java.util.List;
 import java.util.Map;
+import org.fogbowcloud.saps.engine.core.scheduler.executor.arrebol.dtos.CommandResponseDTO;
 
 public class TaskSpecResponseDTO implements Serializable {
 	/**
